+++ tests/test_context_py38.py
@@ -1,11 +1,11 @@
 import pytest
 
-from asphalt.core import Dependency, inject
+from asphalt.core import inject, resource
 
 
 @pytest.mark.asyncio
 async def test_dependency_injection_posonly_argument():
-    async def injected(foo: int, bar: str = Dependency(), /):
+    async def injected(foo: int, bar: str = resource(), /):
         pass
 
     pytest.raises(TypeError, inject, injected).match(