+++ docs/versionhistory.rst
@@ -7,6 +7,9 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 - ``Context`` now accepts parametrized generic classes as resource types
 - Deprecated context variables in favor of dependency injection
+- Replaced ``asphalt.core.context.Dependency`` with
+  ``asphalt.core.context.resource`` due to issues with strict type checking (the former
+  is now deprecated and will be removed in v5.0)
 
 **4.7.0** (2022-04-08)